Implemented basic framework for the Nines Project.
Created a base and shuffled version of the array containing the numbers 1-9.

Created functions to interact with said array: printArray, shuffleArray, threebythree. crossSum.

Added visualization as console outpot for both arrays, a 3 by 3, each row, column and diagonal. As well as each respective cross sum.

Code is still dirty and riddled with unused/commented code and missing documentation in the code.