Implemented interaction with the code to play the nines game.
Implemented functions using prompts to interact with the code.

The user now has the ability to play the nines game by inputting which numbers on the board to reveal, and then to choose which set of three to take as their final sum.

The function gameStart starts the game by going through the reveal and selection process implemented in the functions revealing and setSelect.

The code is still very rough and needs to be refactored and optimized by cutting down reptitive code through other functions.