Implemented a primitive Solving algorithm
Implemented the function solverPrimitive which checks and calculates the optimal choice for choosing the lock in set. It currently only works for sets with already 2 unlocked spots and is being worked on to work for more cases. It functions after every reveal, to also helps making the next choice easier.