Bugfixing and implementation for more cases
Fixed bugs in the primitive solving function that made it so certain elements were not being shown.

Added more cases, so 1,2,3 and 7,8,9 sets are always respected as a possibility. They show with an average value of 361 and 1001, to notify the user about the possibility of a big hit.

Still missing implementation for choosing the best elements to reveal.