Implemented primitive function for slot values
Implemented algorithms to give the user feedback on the value of each slot on the board for their next reveal choice.

Now the user is able to always pick the highest value slot to reveal as well as picking the set with the highest expected value at the end.

The function still needs a way to remove already revealed slots from the list.