Fixed the listing of slotvalues, implemented option to test specific cases
Listing of the slotvalues in function trimOptions fixed: now only shows unrevealed slots sorted by their values descending. It only shows them while the user still is able to reveal further slots.

Added (commented out in a normal run of the code) the option to run a specific set of 9 numbers to test cases for consitency.