Self-playing, Log cleanup, function-Adjustments
Updates:

Cleaned up and added a few more explanations to functions and console-logs.

Adjusted solverPrimitive function:
- Logs can now be toggled based on flag argument
- can now return either the best set or slot choice based on argument input

Self-playing:
The program can now play itself. It will always pick the highest value slot to reveal and then the highest win-value set. This will be useful to simulate the game and gather data.

Added a console-log to showcase the average value of a given board - this allows a comparison between how well the program is doing on any given board configuration, based on its result.