Added functionality for fully math based solving
Added a function solverMathy which now uses actual math to calculate all possible results in a given set of three numbers, regardless of how many slots are revealed or not. This deprecates the solverPrimitive function which operated on only assigning values for the highest possible prizes if 0 or 1 slots were revealed. This makes the solving algorithm more accurate and able to account for mid range prizes that are much better than average/low prizes.