Adjusted many functions for better readability and
changes:

- function threebythree:
usage of template literals and goign by array lenghth instead of static numbers

- function threebythreeNines:
swapped to using a set for the visible slots (faster calculation)
used map approach to decide between highlighted number or X
made it more readable by using arrow function and shortening the if statement by using ternary operator

- function threebythreeFinal
similar to threebythreeNines: swapped to an array passed to threebythree approach along with mapping and an easier readable if to decide between colours. Colouring abstracted from the printing.

- function slotsFromName
swapped switch case to slots object and just extract it from there

- function crossSum
swapped for loop with declarations to using the JavaScript reduce function instead to use sum (0) as an accumulator that gets all numbers added onto itself from the array

- function prizeSelect
adapted the code to use JavaScripts built in findIndex on the prize and check for the sum to find the index value in the prizes array

Lessons learned:

Sets, map(), template literals, switch case swap to a constant object