OOP Adapted, automatic playthrough and game history logging made possible
- Added function analyzeGamehistory
This function aims to calculate all the interesting information from the game histories of games played, to have an overview of the results.

- added function trimOptions
This functions cuts down the best slot choice to only slots that arent revealed yet.

- added function: revealAuto
This function automatically reveals a slot based on the input. (Aimed to be used with solverMath).

added function: chooseSetAuto:
Same as reveal, but for the set choice.

added function: gameStartAuto:
Goes through a game simulation by doing 3 automatic reveals and the set choice, based on the calculation from solverMath.

- adapted getGameResult:
Added all relevant information of a playthrough to be returned as a game history later on.

- added function: simulateGames:
This function lets us play through a self assigned number of simulations and log all the game history to be analyzed afterwards.

Bonus: measureExecutionTime helps in finding out how long the simulations took.